Move HtmlCleaner and subclasses into own file
